feat(examples): add array bracket notation example to jest tests
